adds a radare2 version check
It is known that for some old version this library hangs in a deadlock
probably due to protocol mismatches. We don't really know what is the
lower bound, but we know that at least version 2.3.0 works, so let's
pick it as the first approximation.

The version is checked only once to minimize the number of created
processes. The version parser is made robust enough to withstand
possible variations in the output (including presence of git sha's, and
branch names, and so on).

Since radare supports up to 4.03.0 I had to implement parsing
manually, as `String.split_on` is available only for 4.04.