rewrites IPC with the radare2 process
This PR that resolves #11 and #12 and a few other problems. The error
handling is not ideal but a little bit more robust than it was. That
stderr is no longer hushed and in case if we failed to start the
radare2 process we dump the error message through an exception.